implement async experiments
`the$expeiments` is a list with a number of experiment objects which contain metadata about experiments. They'll eventually have a class and methods.

The `run_experiment()` tool can now optionally be ran asynchrously (and does by default) so that the chat can continue on and multiple experiments can run at once. For now, this is pretty unhelpful as there's no UI to know the status of different experiments. This is next! (#1)